Fix single member selection
This problem only affected the HTML rendering of the tables.

When selecting a single member to collapse the tables (see #6), this was
done before other filters were applied. Thus, further filters could
completely remove a model from the list. E.g. the most evaluated
realization of CanESM5 is r1i1p1f1, which is not available on ESGF.
Before this fix, this realization was the selected member and then
it was eliminated from the list by the filter_avail.

The behaviour has been corrected to take the filter to be applied into
account when selecting the most "representative" member for the table.