Updates of yaml files via Samuel
Original comment:

Qasmi.yaml updated.
It now includes 40 GCMs.
I have put first the performance criteria which is the capacity to fit
or not in a constrained future climate change range Second, I put a
"futur-spread" criteria based on warming classes.  Note that the
performance criteria by Qasmi is very strict (follow-up of Ribes et al.
2021 and article in review) and it eliminates a lot of GCMs.  This is
due to the fact that observational constraints give a very strong
constraint on 2050 climate change values, stronger than on TCR for
example Good to keep it somewhere as an option but not sure we want to
activate it for the GCM selection.

Updated version of the Sevault.yaml.
Very minor changes. Mostly a final check of the values and comments.
Nothing major.