Add EC-Earth aerosol treatment / Update tables EUR
Shuting Yang provided info on aerosol treatment for the EC-Earth
consortium simulations for CMIP6:

"EC-Earth3 and EC-Earth3-Veg use the simple plume approach of
MACv2-SP (Stevens et al., 2017), which provides a simplified, parametric
representation of the optical properties of the anthropogenic
contribution to the tropospheric aerosol burden that are consistent with
the CMIP6 anthropogenic aerosol emission. Only the EC-Earth3-AerChem
configuration that includes an interactive aerosols and atmospheric
chemistry model and simulates the tropospheric aerosol and chemistry has
the output of the aerosol properties. But none of the EC-Earth3-AerChem
simulations saved the model level output for CORDEX."