Fix CESM2 member / Update all for EUR
Fixes #15

CESM2 r2i1p1f1 appeared in many tables as available when the historical
data for driving an RCM are not on ESGF (the scenarios are there,
though). The availability synthesis has been updated to include the check
for the historical run.

All EUR tables have been updated accordingly, including the community selected
GCMs in CMIP6_downscaling_plans.csv