Add skip_disabled_domain option / Change to ssp370
This allows to disable entries for specific domains, even if they meet
the spatial scope criteria. This is one of the options to be able to
handle all domains in the main branch and keep track of all decisions
within the yaml files for the metrics. See #5

In order preserve the EUR table, some disable_EUR entries have been
added to the AtlasIPCC.yaml file.

Note that future spread assessment has been switched to SSP3-7.0, in
accordance with Med-CORDEX decisions. This also makes sense for
EURO-CORDEX since this is a prioritized scenario, unlike SSP5-8.5 that
appeared before in the table trying to maximize spread. There should not
be major changes in the relative changes across models.